Improve initial audio output route setting
On service startup a call to setForceUse(mIsSpeakerUsed) is made to set
the audio route to the built-in speaker (this is by default). Later,
another call is made when FM device is powered up, but if the app was
started with headphones/headset plugged in and the FM device was already
powered up (e.g., when app was force killed through settings and we had
no chance to power down properly), the second call to setForceUse won't
be made to switch to headphones/headset, so the speaker remains the
default audio output, requiring the user to plug out/in the
headphones/headset or power down/up the FM radio.

Signed-off-by: iusmac <iusico.maxim@libero.it>